move proposals to base and add multivariate cauchy
